refactor: migrate icon handling to new Icon component and update API query keys for consistency
