+++ web/src/components/SignInCard.tsx
@@ -11,7 +11,7 @@ import { TextInput } from '@astryxdesign/core/TextInput';
 import { Controller, useForm, useWatch } from 'react-hook-form';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useToast } from '@/lib/hooks/use-toast';
-import { apiQueryKey, requestApiJson } from '@/lib/api';
+import { requestApiJson } from '@/lib/api';
 import { WelcomeTitle } from '@/components/WelcomeTitle';
 
 const loginSchema = z.object({
@@ -41,7 +41,7 @@ export function SignInCard({ initialEmail = '' }: { initialEmail?: string }) {
     async function handlePasswordSignIn(payload: LoginValues) {
         try {
             await login.mutateAsync(payload);
-            await queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/me') });
+            await queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/me'] });
             navigate('/organizations', { replace: true });
         } catch (loginError) {
             showToast({
