+++ web/src/components/dialogs/CreateCompute.tsx
@@ -11,7 +11,7 @@ import { Dialog, DialogHeader } from '@astryxdesign/core/Dialog';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Layout, LayoutContent, LayoutFooter } from '@astryxdesign/core/Layout';
 import { useToast } from '@/lib/hooks/use-toast';
-import { apiQueryKey, fetchApiJson } from '@/lib/api';
+import { fetchApiJson } from '@/lib/api';
 import { zComputeRegistryResponse } from '@/lib/generated/platform-api-v1/zod.gen';
 
 const schema = z.object({
@@ -44,7 +44,7 @@ export default function CreateCompute() {
         onSuccess: () => {
             setOpen(false);
             form.reset();
-            return queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/computes') });
+            return queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/computes'] });
         },
     });
