+++ web/src/components/dialogs/CreateDatabase.tsx
@@ -13,7 +13,7 @@ import { Dialog, DialogHeader } from '@astryxdesign/core/Dialog';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Layout, LayoutContent, LayoutFooter } from '@astryxdesign/core/Layout';
 import { useToast } from '@/lib/hooks/use-toast';
-import { apiQueryKey, fetchApiJson } from '@/lib/api';
+import { fetchApiJson } from '@/lib/api';
 import { zDatabaseRegistryResponse, zDatabaseSslMode } from '@/lib/generated/platform-api-v1/zod.gen';
 
 const schema = z.object({
@@ -52,7 +52,7 @@ export default function CreateDatabase() {
         onSuccess: () => {
             setOpen(false);
             form.reset();
-            return queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/databases') });
+            return queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/databases'] });
         },
     });
