+++ web/src/components/dialogs/CreateOrganization.tsx
@@ -10,7 +10,7 @@ import { Dialog, DialogHeader } from '@astryxdesign/core/Dialog';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Layout, LayoutContent, LayoutFooter } from '@astryxdesign/core/Layout';
 import { useToast } from '@/lib/hooks/use-toast';
-import { apiQueryKey, requestApi } from '@/lib/api';
+import { requestApi } from '@/lib/api';
 
 const createOrganizationSchema = z.object({
     name: z.string().trim().min(1),
@@ -35,8 +35,8 @@ export default function CreateOrganization() {
             }),
         onSuccess: () =>
             Promise.all([
-                queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/organizations') }),
-                queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/me/organizations') }),
+                queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/organizations'] }),
+                queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/me/organizations'] }),
             ]),
     });
     const formId = useId();
