+++ web/src/components/dialogs/CreateStorage.tsx
@@ -10,7 +10,7 @@ import { Dialog, DialogHeader } from '@astryxdesign/core/Dialog';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Layout, LayoutContent, LayoutFooter } from '@astryxdesign/core/Layout';
 import { useToast } from '@/lib/hooks/use-toast';
-import { apiQueryKey, fetchApiJson } from '@/lib/api';
+import { fetchApiJson } from '@/lib/api';
 import { zStorageRegistryResponse } from '@/lib/generated/platform-api-v1/zod.gen';
 
 const schema = z.object({
@@ -50,7 +50,7 @@ export default function CreateStorage() {
         onSuccess: () => {
             setOpen(false);
             form.reset();
-            return queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/storages') });
+            return queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/storages'] });
         },
     });
