+++ web/src/components/ui/Tabs.tsx
@@ -1,8 +1,7 @@
 import { Stack } from '@astryxdesign/core/Stack';
 import { Tab as AstryxTab, TabList as AstryxTabList } from '@astryxdesign/core/TabList';
 import { useState, Children, isValidElement, type ComponentProps, type ReactElement, type ReactNode } from 'react';
-import type { StoneIconName } from '@/icons';
-import { Icon } from '@/components/ui/Icon';
+import { Icon, type StoneIconName } from '@/components/ui/Icon';
 
 type TabsProps = Omit<ComponentProps<typeof AstryxTabList>, 'children' | 'onChange' | 'value'> & {
     children?: ReactNode;
