--- web/src/icons.tsx
@@ -1,81 +1,0 @@
-import type { LucideIcon } from 'lucide-react';
-import type { IconRegistry } from '@astryxdesign/core/Icon';
-import {
-    X,
-    ChevronDown,
-    ChevronLeft,
-    ChevronRight,
-    ChevronsLeft,
-    ChevronsRight,
-    Check,
-    CheckCircle,
-    XCircle,
-    AlertTriangle,
-    Info,
-    Calendar,
-    Clock,
-    ExternalLink,
-    Menu,
-    MoreHorizontal,
-    Search,
-    ArrowUp,
-    ArrowDown,
-    ArrowUpDown,
-    Boxes,
-    Building2,
-    Database,
-    Filter,
-    HardDrive,
-    EyeOff,
-    Columns,
-    Copy,
-    CheckCheck,
-    Wrench,
-    Square,
-    UserRound,
-    Users,
-    Mic,
-} from 'lucide-react';
-
-export const stoneIconComponents = {
-    close: X,
-    chevronDown: ChevronDown,
-    chevronLeft: ChevronLeft,
-    chevronRight: ChevronRight,
-    chevronsLeft: ChevronsLeft,
-    chevronsRight: ChevronsRight,
-    check: Check,
-    success: CheckCircle,
-    error: XCircle,
-    warning: AlertTriangle,
-    info: Info,
-    calendar: Calendar,
-    clock: Clock,
-    externalLink: ExternalLink,
-    menu: Menu,
-    moreHorizontal: MoreHorizontal,
-    search: Search,
-    arrowUp: ArrowUp,
-    arrowDown: ArrowDown,
-    arrowsUpDown: ArrowUpDown,
-    boxes: Boxes,
-    building2: Building2,
-    database: Database,
-    funnel: Filter,
-    eyeSlash: EyeOff,
-    viewColumns: Columns,
-    copy: Copy,
-    checkDouble: CheckCheck,
-    wrench: Wrench,
-    stop: Square,
-    microphone: Mic,
-    hardDrive: HardDrive,
-    userRound: UserRound,
-    users: Users,
-} satisfies Record<string, LucideIcon>;
-
-export type StoneIconName = keyof typeof stoneIconComponents;
-
-export const stoneIconRegistry: IconRegistry = Object.fromEntries(
-    Object.entries(stoneIconComponents).map(([name, Icon]) => [name, <Icon aria-hidden={true} key={name} size="1em" />])
-) as IconRegistry;
