+++ web/src/lib/api.ts
@@ -1,7 +1,3 @@
-import { hasProtocol } from 'ufo';
-
-const apiBaseUrl = import.meta.env.VITE_API_URL || '';
-
 export type ApiQueryKey = readonly ['api', string];
 
 /** Error thrown for failed API responses. */
@@ -15,43 +11,6 @@ export class ApiError extends Error {
     }
 }
 
-/** Builds the canonical React Query key for one API resource. */
-export function apiQueryKey(path: string): ApiQueryKey {
-    return ['api', apiUrl(path)];
-}
-
-/** Resolves an API path against the configured API origin. */
-function apiUrl(path: string): string {
-    // Reject path separators that could bypass URL checks.
-    if (path.includes('\\')) {
-        throw new Error('API path must not contain backslashes');
-    }
-
-    // Validate absolute API URLs before using them.
-    if (hasProtocol(path)) {
-        const url = new URL(path);
-
-        // Only browser HTTP(S) URLs are allowed.
-        if (url.protocol !== 'http:' && url.protocol !== 'https:') {
-            throw new Error('API URL must use HTTP(S)');
-        }
-
-        return path;
-    }
-
-    // Reject protocol-relative URLs.
-    if (path.startsWith('//')) {
-        throw new Error('API path must be relative or absolute HTTP(S)');
-    }
-
-    // Keep relative paths unchanged when no API origin is configured.
-    if (!apiBaseUrl) {
-        return path;
-    }
-
-    return new URL(path, apiBaseUrl).toString();
-}
-
 /** Sends one API request and normalizes non-OK errors. */
 export async function requestApi(path: string, init?: RequestInit, fetchImpl: typeof fetch = fetch): Promise<Response> {
     const headers = new Headers(init?.headers);
@@ -61,7 +20,7 @@ export async function requestApi(path: string, init?: RequestInit, fetchImpl: ty
         headers.set('Accept', 'application/json');
     }
 
-    const response = await fetchImpl(apiUrl(path), {
+    const response = await fetchImpl(path, {
         ...init,
         credentials: 'include',
         headers,
