+++ web/src/platform/layouts/Application.tsx
@@ -8,8 +8,8 @@ import { TopNav } from '@astryxdesign/core/TopNav';
 import { AppShell } from '@astryxdesign/core/AppShell';
 import { Tab, TabList } from '@astryxdesign/core/TabList';
 import { findActiveTab } from '@/lib/paths';
-import { getIconComponent } from '@/lib/icons';
 import { Wordmark } from '@/components/Wordmark';
+import { getIconComponent } from '@/components/ui/Icon';
 import { PageContainer } from '@/components/PageContainer';
 import { DevelopmentNotice } from '@/components/DevelopmentNotice';
 import { pageRouteIsDynamic, type RuntimePage } from '@/platform/pages';
