+++ web/src/platform/routes/admin/Compute.tsx
@@ -12,7 +12,7 @@ import type { ComputeRegistryResponse } from '@/lib/generated/platform-api-v1/ty
 import { useDeleteDialog } from '@/lib/utils';
 import { useToast } from '@/lib/hooks/use-toast';
 import { useApiQuery } from '@/lib/hooks/use-api';
-import { apiQueryKey, requestApi } from '@/lib/api';
+import { requestApi } from '@/lib/api';
 import { usePaginate } from '@/lib/hooks/pagination';
 import { Table, TableColumn } from '@/components/ui/Table';
 import CreateCompute from '@/components/dialogs/CreateCompute';
@@ -28,7 +28,7 @@ export default function AdminCompute() {
             await requestApi(`/api/v1/computes/${computeId}`, { method: 'DELETE' });
         },
         onSuccess: async () => {
-            await queryClient.invalidateQueries({ queryKey: apiQueryKey('/api/v1/computes') });
+            await queryClient.invalidateQueries({ queryKey: ['api', '/api/v1/computes'] });
             toast({ body: 'Compute deleted' });
         },
     });
